+++ R/predict.R
@@ -252,6 +252,9 @@ predict.ranger.forest <- function(object, data, predict.all = FALSE,
   regularization.usedepth <- FALSE
   confounders <- matrix(c(0, 0))
   use.confounders <- FALSE
+  node.stats <- FALSE
+  time.interest <- c(0, 0)
+  use.time.interest <- FALSE
   
   ## Use sparse matrix
   if (inherits(x, "dgCMatrix")) {
@@ -276,7 +279,7 @@ predict.ranger.forest <- function(object, data, predict.all = FALSE,
                       prediction.type, num.random.splits, sparse.x, use.sparse.data,
                       order.snps, oob.error, max.depth, inbag, use.inbag, 
                       regularization.factor, use.regularization.factor, regularization.usedepth, 
-                      confounders, use.confounders)
+                      node.stats, time.interest, use.time.interest, confounders, use.confounders)
 
   if (length(result) == 0) {
     stop("User interrupt or internal error.")