+++ NEWS.md
@@ -1,6 +1,13 @@
 
-# ranger 0.15.4
+# ranger 0.16.1
 * Add linear model residual splitting
+
+# ranger 0.16.0
+* New CRAN version
+
+# ranger 0.15.4
+* Add node.stats option to save node statistics of all nodes
+* Add time.interest option to restrict unique survival times (faster and saves memory)
 
 # ranger 0.15.3
 * Fix min bucket option in C++ version